feat: new blog left panel layout
