+++ gatsby-node.ts
@@ -14,7 +14,7 @@ const path = require(`path`)
 const { createFilePath } = require(`gatsby-source-filesystem`)
 
 // Define the template for blog and blog post
-const blogPost = path.resolve(`./src/templates/blog-post.tsx`)
+const blogPost = path.resolve(`./src/templates/blog-post/index.tsx`)
 const blog = path.resolve(`./src/templates/blog.tsx`)
 const comparisonTemplate = path.resolve(`./src/templates/product-comparison.tsx`)
