+++ src/style.less
@@ -5852,7 +5852,7 @@ table th {
 /*Why Estuary end*/
 /*Blog post template updates*/
 .blog-post{
-    max-width: 1220px;
+    max-width: 1448px;
     width: 100%;
     padding: 2rem 0 5rem;
     @media(max-width: 1284px){
@@ -5945,6 +5945,7 @@ table th {
         display: flex;
         align-items: flex-start;
         justify-content: space-between;
+        gap: 110px;
         padding-top: 48px;
         margin-top: 24px;
         border-top: 1px solid #EEF0F3;
@@ -6020,9 +6021,8 @@ table th {
         }
     }
     .post-sidebar{
-        max-width: 422px;
+        max-width: 215px;
         width: 100%;
-        padding-right: 40px;
         @media(max-width: 991px){
             max-width: 322px;
         }
@@ -6145,7 +6145,7 @@ table th {
     .table-of-contents{
         background: #F9FAFC;
         counter-reset: toc-counter;
-        padding: 16px;
+
         h3{
             margin: 0 0 8px;
             color: #04192B;
@@ -6156,9 +6156,23 @@ table th {
         }
         ul{
             margin: 0;
+            padding: 0 16px;
+
             li{
-                margin-bottom: 8px;
                 list-style: none;
+                opacity: 0.7;
+                position: relative;
+                margin: 0;
+
+                & .progress {
+                    background-color: #D9D9D9;
+                    position: absolute;
+                    top: 0;
+                    bottom: 0;
+                    right: -16px;
+                    width: 3px;
+                }
+
                 &:not(.sub-item) {
                     counter-increment: list-number;
                     a {
@@ -6181,20 +6195,39 @@ table th {
                     }
                 }
                 a{
-                    color: #47506D;
+                    color: #333F51;
                     font-size: 14px;
                     font-style: normal;
-                    font-weight: 500;
+                    font-weight: 400;
                     line-height: 22px;
                     &:hover{
                         text-decoration: underline;
                     }
+                }
+                &.active {
+                    opacity: 1;
+                    & .progress {
+                        background-color: #333F51;
+
+                        &.current:after {
+                            content: " \25CF";
+                            width: 12px;
+                            height: 12px;
+                            font-size: 12px;
+                            position: absolute;
+                            bottom: 0;
+                            left: -4px;
+                            z-index: 2;
+                        }
+                    }
+
                 }
                 &.sub-item{
                     list-style: disc;
                     margin-left: 50px;
                     font-weight: 400;
                     line-height: 18px;
+                    font-size: 12px;
                     span{
                         display: none;
                     }
